no route handling more consistent, overflow in faktorybauer avoided by extending array_tpl to 32 bit, sign rotation
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@1365 8aca7d54-2c30-db11-9de9-000461428c89

+++ tpl/array_tpl.h
@@ -13,7 +13,7 @@
 template<class T> class array_tpl
 {
 	public:
-		typedef uint16 index;
+		typedef uint32 index;
 
 		explicit array_tpl() : data(NULL), size(0) {}
